requirements.txt
keboola.component==1.4.0
keboola.utils
keboola.http-client
mock
freezegun
xmltodict
